+++ packages/af-utils-react-virtual-headless/src/hooks/useOnce/index.js
@@ -1,5 +1,0 @@
-import { useRef } from "react";
-
-const useOnce = cb => (useRef().current ||= cb());
-
-export default useOnce;