+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtualModel/index.js
@@ -1,17 +1,15 @@
+import { useRef } from "react";
 import List from "models/List";
 import { DEFAULT_ESTIMATED_WIDGET_SIZE } from "constants";
 import mergeModelParams from "/utils/mergeModelParams";
 
-import useOnce from "../useOnce";
+const useVirtualModel = params => {
+    const model = (useRef().current ||= new List(
+        params.estimatedWidgetSize ?? DEFAULT_ESTIMATED_WIDGET_SIZE
+    ));
+    mergeModelParams(model, params);
 
-const useVirtualModel = params =>
-    useOnce(() => {
-        const model = new List(
-            params.estimatedWidgetSize ?? DEFAULT_ESTIMATED_WIDGET_SIZE
-        );
-        mergeModelParams(model, params);
-
-        return model;
-    });
+    return model;
+};
 
 export default useVirtualModel;