+++ packages/af-utils-react-virtual-headless/src/index.js
@@ -4,7 +4,6 @@ export { default as useVirtual } from "./hooks/useVirtual";
 export { default as useVirtualModel } from "./hooks/useVirtualModel";
 export { default as useSubscription } from "./hooks/useSubscription";
 export { default as useComponentSubscription } from "./hooks/useComponentSubscription";
-export { default as useOnce } from "./hooks/useOnce";
 
 export { default as Subscription } from "./components/Subscription";