+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @af-utils/react-mobx-table
 
+## 0.0.50
+
+### Patch Changes
+
+-   removed useOnce export
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.42
+    -   @af-utils/react-table@0.0.48
+
 ## 0.0.49
 
 ### Patch Changes